Changes grid spacing to an integral type and adjusts UI elements to match the new spacing.

+++ EXTChartView.h
@@ -41,7 +41,7 @@
 + (CGRect)dotBoundingBoxForCellRank:(NSInteger)cellRank
                           termIndex:(NSInteger)termIndex
                        gridLocation:(EXTIntPoint)gridLocation
-                        gridSpacing:(CGFloat)gridSpacing;
+                        gridSpacing:(NSInteger)gridSpacing;
 @end
